Move querying for sensor info to ContextManager. Add authority to vdb sensors.
This change set moves fetching of sensor information to the ContextManager
since many activities are going to want to find out what sensors are
available, including the Expression Builder.

This change also adds authorities to the meta data for sensors for
those that are VDB based sensors so that we can tell that they
are VDB based when we query for them.