Significant refactoring of the expression system to enable persistence.
The goal of this change set is to make it easy to persist expressions
in a future change set. This changeset introduces a number of enums
designed to make all constants in the expression exist in only one place
and to enable switching on various constants for more readable code.

It also addresses a few issues with Parceling TypedValues which were
found via checkstyle warnings.

We also added a BOOT broadcast receiver which currently does nothing
but which is intended to reload persisted expressions so that context
droid can begin running properly after a reboot.