Add expression parsing using ANTLRv3.
This changeset cleans up the expressions even further and adds expression
parsing using a parser generated by ANTLRv3 and a custom grammar for
context expressions. The grammar could stand to be improved to be
more flexible in terms of parenthesis and order of operations
for boolean operations but it is good enough for parsing persisted
expressions.