Major reworking of the grammar and expressions for presidence.
This changeset involves a large number of changes to the expression
grammar in order to support presidence properly. This also required
some refactoring of the Expression classes themselves in order
to have a consistent type for all expressions.

The entity value path separator also changes as does the internal
value path separator, though I am not sure we are using that anywhere
yet. This was required because of ambiguities with value paths
and division and I didn't want to have whitespace have
syntactic value and have to deal with whitespace everywhere
in the grammar.

Further work is still needed as it would be good to get rid of
the "expression/value" dichotomy since we have to wrap values
in the (new and probably temporary) TypedValueExpression to
wrap values into something that can be used in arbitrary
positions within an expression.

Further work is still needed on the test suite to make sure
everything is totally correct.